Add files via upload
Excel Dynamic Dashboard (Beginners)- Data Analytics Project. 

In this project I have used a call dataset from one of the organization. at the top it has 3 KPI’s which list the total calls, the average score for customer satisfaction and the call duration average. 

I have used line chat to represent the trend of call for the particular month, a doughnut chart represents call channel. Then we have a map , a column chart for the reason the person called and last but not least, a bar chart indicating the response time.

I have used the slicers that can be seen on the right side. We can use them to filter our data and thus make our dashboard